etl/pyproject.toml
[project]
name = "fusiondex-etl"
version = "0.2.0"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32",
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0",
    "scrapy>=2.11",
    "pillow>=11.0",
    "prefect>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["etl"]

[tool.uv]
dev-dependencies = []
